Abstracted the state into its own class
