+++ fto3phase/src/main/java/cs/fto3phase/FullFto.java
@@ -8,20 +8,20 @@ public class FullFto {
     /**
      * Internal State
      */
-    int corners = SOLVED_CORNERS;
-    long edges = SOLVED_EDGES;
-    long centers = SOLVED_CENTERS;
-    long packedCenterIndicesLow = SOLVED_CENTER_INDICES_LOW;
-    long packedCenterIndicesHigh = SOLVED_CENTER_INDICES_HIGH;
+
+    private InnerState state = new InnerState();
+    private InnerState rot1 = new InnerState();
+    private InnerState rot2 = new InnerState();
+
+
     private boolean trackCenterIndices;
     private static boolean trackCenterIndicesByDefault = true;
 
     /**
      * History, used for undo() method
      */
-    private Stack<Integer> cornerHistory = new Stack<Integer>();
-    private Stack<Long> edgeHistory = new Stack<Long>();
-    private Stack<Long> centerHistory = new Stack<Long>();
+    private Stack<InnerState> stateHistory = new Stack<InnerState>();
+
     private Stack<Long> centerIndicesLowHistory = new Stack<Long>();
     private Stack<Long> centerIndicesHighHistory = new Stack<Long>();
     private Stack<Move> moveHistory = new Stack<Move>();
@@ -42,96 +42,15 @@ public class FullFto {
      * @param fto copy from
      */
     public FullFto(FullFto fto){
-        this.corners = fto.corners;
-        this.edges = fto.edges;
-        this.centers = fto.centers;
-        this.packedCenterIndicesLow = fto.packedCenterIndicesLow;
-        this.packedCenterIndicesHigh = fto.packedCenterIndicesHigh;
+        this.state.corners = fto.state.corners;
+        this.state.edges = fto.state.edges;
+        this.state.centers = fto.state.centers;
         this.trackCenterIndices = fto.trackCenterIndices;
 
-        this.cornerHistory.addAll(fto.cornerHistory);
-        this.edgeHistory.addAll(fto.edgeHistory);
-        this.centerHistory.addAll(fto.centerHistory);
+        this.stateHistory.addAll(fto.stateHistory);
         this.centerIndicesLowHistory.addAll(fto.centerIndicesLowHistory);
         this.centerIndicesHighHistory.addAll(fto.centerIndicesHighHistory);
         this.moveHistory.addAll(fto.moveHistory);
-    }
-
-    /**
-     * Example of what solved corners/edges/centers looks like
-     */
-    private static final int SOLVED_CORNERS;
-    private static final long SOLVED_EDGES;
-    private static final long SOLVED_CENTERS;
-    private static final long SOLVED_CENTER_INDICES_LOW;
-    private static final long SOLVED_CENTER_INDICES_HIGH;
-
-    /**
-     * Initialization of bit-hacking
-     */
-    private static final int CORNER_BITS = 5;
-    private static final int CORNER_MASK = 0b11111;
-    private static final int[] CORNER_MASKS = new int[6];
-    static {
-        for (int i = 0; i < 6; i++) {
-            CORNER_MASKS[i] = CORNER_MASK << (CORNER_BITS * i);
-        }
-
-        int solvedCorners = 0;
-        for (int i = 0; i < 6; i++) {
-            solvedCorners |= encodeCorner(i, 0) << (CORNER_BITS * i);
-        }
-        SOLVED_CORNERS = solvedCorners;
-    }
-
-    private static final int EDGE_BITS = 4;
-    private static final long EDGE_MASK = 0b1111;
-    private static final long[] EDGE_MASKS = new long[12];
-    static {
-        for (int i = 0; i < 12; i++) {
-            EDGE_MASKS[i] = EDGE_MASK << (EDGE_BITS * i);
-        }
-
-        long solvedEdges = 0;
-        for (long i = 0; i < 12; i++) {
-            solvedEdges |= i << (EDGE_BITS * i);
-        }
-        SOLVED_EDGES = solvedEdges;
-    }
-
-    private static final int CENTER_BITS = 2;
-    private static final long CENTER_MASK = 0b11;
-    private static final long[] CENTER_MASKS = new long[24];
-    static {
-        for (int i = 0; i < 24; i++) {
-            CENTER_MASKS[i] = CENTER_MASK << (CENTER_BITS * i);
-        }
-
-        long solvedCenters = 0;
-        for (int i = 0; i < 8; i++) {
-            for (int j = 0; j < 3; j++){
-                solvedCenters |= (long) (i & CENTER_MASK) << (CENTER_BITS * ((i * 3) + j));
-            }
-        }
-        SOLVED_CENTERS = solvedCenters;
-    }
-
-    private static final int CENTER_INDEX_BITS = 5;
-    private static final long CENTER_INDEX_MASK = 0b11111;
-    private static final long[] CENTER_INDEX_MASKS = new long[12];
-    static {
-        for (int i = 0; i < 12; i++) {
-            CENTER_INDEX_MASKS[i] = CENTER_INDEX_MASK << (CENTER_INDEX_BITS * i);
-        }
-
-        long solvedCenterIndicesLow = 0;
-        long solvedCenterIndicesHigh = 0;
-        for (int i = 0; i < 12; i++) {
-            solvedCenterIndicesLow |= (long) i << (CENTER_INDEX_BITS * i);
-            solvedCenterIndicesHigh |= (long) (i + 12) << (CENTER_INDEX_BITS * i);
-        }
-        SOLVED_CENTER_INDICES_LOW = solvedCenterIndicesLow;
-        SOLVED_CENTER_INDICES_HIGH = solvedCenterIndicesHigh;
     }
 
     /**
@@ -139,11 +58,7 @@ public class FullFto {
      * @return number of moves
      */
     public int historyLength() {
-        assert (centerHistory.size() == edgeHistory.size());
-        assert (edgeHistory.size() == cornerHistory.size());
-        assert (!trackCenterIndices || centerIndicesLowHistory.size() == centerHistory.size());
-        assert (!trackCenterIndices || centerIndicesHighHistory.size() == centerHistory.size());
-        assert (edgeHistory.size() == moveHistory.size());
+        assert (moveHistory.size() == stateHistory.size());
         return moveHistory.size();
     }
 
@@ -151,12 +66,10 @@ public class FullFto {
      * Get rid of history. undo() and isRepetition() are the effected methods.
      */
     public void clearMoveStack(){
-        centerHistory.clear();
         centerIndicesLowHistory.clear();
         centerIndicesHighHistory.clear();
-        edgeHistory.clear();
-        cornerHistory.clear();
         moveHistory.clear();
+        stateHistory.clear();
     }
 
     /**
@@ -204,108 +117,6 @@ public class FullFto {
     }
 
     //------------------BITWISE HELPER OPERATIONS------------------//
-    private static int encodeCorner(int perm, int orientation){
-        return ((perm << 2) | orientation);
-    }
-
-    private static int getCornerIndex(int corner){
-        return corner >> 2;
-    }
-
-    private static int getCornerOrientation(int corner){
-        return corner & 0b11;
-    }
-
-    private int getCorner(int i){
-        return (corners >> (CORNER_BITS * i)) & CORNER_MASK;
-    }
-
-    private void setCorner(int i, int corner){
-        int shift = CORNER_BITS * i;
-        corners = (corners & ~CORNER_MASKS[i]) | (corner << shift);
-    }
-
-    private void cycleCorners(int i1, int i2, int i3){
-        int c1 = getCorner(i1);
-        int c2 = getCorner(i2);
-        int c3 = getCorner(i3);
-
-        setCorner(i1, c3);
-        setCorner(i2, c1);
-        setCorner(i3, c2);
-    }
-
-    private void swapCorners(int i1, int i2){
-        int c1 = getCorner(i1);
-        int c2 = getCorner(i2);
-
-        setCorner(i1, c2);
-        setCorner(i2, c1);
-    }
-
-    private void twistCorner(int i, int dir){
-        int corner = getCorner(i);
-        setCorner(i, encodeCorner(getCornerIndex(corner), (getCornerOrientation(corner) + dir) & 0b11));
-    }
-
-    private int getEdge(int i){
-        return (int) ((edges >> (EDGE_BITS * i)) & EDGE_MASK);
-    }
-
-    private void setEdge(int i, int edge){
-        int shift = EDGE_BITS * i;
-        edges = (edges & ~EDGE_MASKS[i]) | ((long) edge << shift);
-    }
-
-    private void cycleEdges(int i1, int i2, int i3){
-        int e1 = getEdge(i1);
-        int e2 = getEdge(i2);
-        int e3 = getEdge(i3);
-
-        setEdge(i1, e3);
-        setEdge(i2, e1);
-        setEdge(i3, e2);
-    }
-
-    private void swapEdges(int i1, int i2){
-        int e1 = getEdge(i1);
-        int e2 = getEdge(i2);
-
-        setEdge(i1, e2);
-        setEdge(i2, e1);
-    }
-
-    private int getCenter(int i){
-        return (int) ((centers >> (CENTER_BITS * i)) & CENTER_MASK);
-    }
-
-    private void setCenter(int i, int center){
-        int shift = CENTER_BITS * i;
-        centers = (centers & ~CENTER_MASKS[i]) | ((long) center << shift);
-    }
-
-    /**
-     * Get center index in centerIndices
-     * @param i
-     * @return
-     */
-    private int getCenterIndex(int i){
-        ensureCenterIndexTracking();
-        int packedIndex = i % 12;
-        long packed = i < 12 ? packedCenterIndicesLow : packedCenterIndicesHigh;
-        return (int) ((packed >> (CENTER_INDEX_BITS * packedIndex)) & CENTER_INDEX_MASK);
-    }
-
-    private void setCenterIndex(int i, int centerIndex){
-        ensureCenterIndexTracking();
-        int packedIndex = i % 12;
-        int shift = CENTER_INDEX_BITS * packedIndex;
-        if (i < 12) {
-            packedCenterIndicesLow = (packedCenterIndicesLow & ~CENTER_INDEX_MASKS[packedIndex]) | ((long) centerIndex << shift);
-        } else {
-            packedCenterIndicesHigh = (packedCenterIndicesHigh & ~CENTER_INDEX_MASKS[packedIndex]) | ((long) centerIndex << shift);
-        }
-    }
 
     private void ensureCenterIndexTracking() {
         if (!trackCenterIndices) {
@@ -313,43 +124,7 @@ public class FullFto {
         }
     }
 
-    private void cycleThreeCenters(int i1, int i2, int i3){
-        int c1 = getCenter(i1);
-        int c2 = getCenter(i2);
-        int c3 = getCenter(i3);
 
-        setCenter(i1, c3);
-        setCenter(i2, c1);
-        setCenter(i3, c2);
-
-        if (!trackCenterIndices) {
-            return;
-        }
-
-        int ci1 = getCenterIndex(i1);
-        int ci2 = getCenterIndex(i2);
-        int ci3 = getCenterIndex(i3);
-        setCenterIndex(i1, ci3);
-        setCenterIndex(i2, ci1);
-        setCenterIndex(i3, ci2);
-    }
-
-    private void swapCenters(int i1, int i2){
-        int c1 = getCenter(i1);
-        int c2 = getCenter(i2);
-
-        setCenter(i1, c2);
-        setCenter(i2, c1);
-
-        if (!trackCenterIndices) {
-            return;
-        }
-
-        int ci1 = getCenterIndex(i1);
-        int ci2 = getCenterIndex(i2);
-        setCenterIndex(i1, ci2);
-        setCenterIndex(i2, ci1);
-    }
 
     /**
      * MATCHING_CENTERS[getCornerIndex(corner)][corner orientation]
@@ -378,9 +153,9 @@ public class FullFto {
     };
 
     private boolean isTriple(int cornerLocation){
-        int corner = getCorner(cornerLocation);
-        int cornerIndex = getCornerIndex(corner);
-        int cornerOrientation = getCornerOrientation(corner);
+        int corner = state.getCorner(cornerLocation);
+        int cornerIndex = state.getCornerIndex(corner);
+        int cornerOrientation = state.getCornerOrientation(corner);
 
         int matchingCenterOne = MATCHING_CENTERS[cornerIndex][cornerOrientation];
         int matchingCenterTwo = MATCHING_CENTERS[cornerIndex][(cornerOrientation + 2) % 4];
@@ -388,14 +163,14 @@ public class FullFto {
         int testSpotOne = TRIPLE_LOCATIONS[cornerLocation][0];
         int testSpotTwo = TRIPLE_LOCATIONS[cornerLocation][1];
 
-        return getCenter(testSpotOne) == matchingCenterOne &&
-                getCenter(testSpotTwo) == matchingCenterTwo;
+        return state.getCenter(testSpotOne) == matchingCenterOne &&
+                state.getCenter(testSpotTwo) == matchingCenterTwo;
     }
 
     public int triplePairsOnCorner(int cornerLocation){
-        int corner = getCorner(cornerLocation);
-        int cornerIndex = getCornerIndex(corner);
-        int cornerOrientation = getCornerOrientation(corner);
+        int corner = state.getCorner(cornerLocation);
+        int cornerIndex = state.getCornerIndex(corner);
+        int cornerOrientation = state.getCornerOrientation(corner);
 
         int matchingCenterOne = MATCHING_CENTERS[cornerIndex][cornerOrientation];
         int matchingCenterTwo = MATCHING_CENTERS[cornerIndex][(cornerOrientation + 2) % 4];
@@ -404,18 +179,18 @@ public class FullFto {
         int testSpotTwo = TRIPLE_LOCATIONS[cornerLocation][1];
 
         int count = 0;
-        if (getCenter(testSpotOne) == matchingCenterOne)
+        if (state.getCenter(testSpotOne) == matchingCenterOne)
             count++;
-        if (getCenter(testSpotTwo) == matchingCenterTwo)
+        if (state.getCenter(testSpotTwo) == matchingCenterTwo)
             count++;
 
         return count;
     }
 
     public int tripleIndexHelper(int cornerLocation){
-        int corner = getCorner(cornerLocation);
-        int cornerIndex = getCornerIndex(corner);
-        int cornerOrientation = getCornerOrientation(corner);
+        int corner = state.getCorner(cornerLocation);
+        int cornerIndex = state.getCornerIndex(corner);
+        int cornerOrientation = state.getCornerOrientation(corner);
 
         int matchingCenterOne = MATCHING_CENTERS[cornerIndex][cornerOrientation];
         int matchingCenterTwo = MATCHING_CENTERS[cornerIndex][(cornerOrientation + 2) % 4];
@@ -424,9 +199,9 @@ public class FullFto {
         int testSpotTwo = TRIPLE_LOCATIONS[cornerLocation][1];
 
         int count = 0;
-        if (getCenter(testSpotOne) == matchingCenterOne)
+        if (state.getCenter(testSpotOne) == matchingCenterOne)
             count |= 1;
-        if (getCenter(testSpotTwo) == matchingCenterTwo)
+        if (state.getCenter(testSpotTwo) == matchingCenterTwo)
             count |= 2;
 
         return count;
@@ -480,7 +255,7 @@ public class FullFto {
         int faceIndex = face.ordinal();
         int[] faceEdges = EDGES_ON_FACE[faceIndex];
         for (int i = 0; i < 3; i++) {
-            if (getEdge(faceEdges[(i + angle) % 3]) != faceEdges[i]){
+            if (state.getEdge(faceEdges[(i + angle) % 3]) != faceEdges[i]){
                 return false;
             }
         }
@@ -511,7 +286,7 @@ public class FullFto {
         int centerId = face.id;
         int[] faceCenters = CENTERS_ON_FACE[face.ordinal()];
         for (int i = 0; i < 3; i++) {
-            if (getCenterOrdinal(faceCenters[i]) != centerId){
+            if (state.getCenterOrdinal(faceCenters[i]) != centerId){
                 return false;
             }
         }
@@ -526,19 +301,19 @@ public class FullFto {
     }
 
     public boolean isPhaseOne(){
-        if (getCenterOrdinal(21) != CenterOrd.D.id ||
-            getCenterOrdinal(22) != CenterOrd.D.id ||
-            getCenterOrdinal(23) != CenterOrd.D.id)
+        if (state.getCenterOrdinal(21) != CenterOrd.D.id ||
+            state.getCenterOrdinal(22) != CenterOrd.D.id ||
+            state.getCenterOrdinal(23) != CenterOrd.D.id)
             return false;
-        return (getEdge(9) == 9 &&
-            getEdge(10) == 10 &&
-            getEdge(11) == 11) ||
-                (getEdge(9) == 11 &&
-                    getEdge(10) == 9 &&
-                    getEdge(11) == 10) ||
-                (getEdge(9) == 10 &&
-                    getEdge(10) == 11 &&
-                    getEdge(11) == 9);
+        return (state.getEdge(9) == 9 &&
+            state.getEdge(10) == 10 &&
+            state.getEdge(11) == 11) ||
+                (state.getEdge(9) == 11 &&
+                    state.getEdge(10) == 9 &&
+                    state.getEdge(11) == 10) ||
+                (state.getEdge(9) == 10 &&
+                    state.getEdge(10) == 11 &&
+                    state.getEdge(11) == 9);
     }
 
     public boolean isPhaseTwo(){
@@ -589,9 +364,7 @@ public class FullFto {
         }
     }
 
-    private int getCenterOrdinal(int index){
-        return (int)(centers >> (2 * index) & CENTER_MASK) + (index > 11 ? 4 : 0);
-    }
+
 
     private static boolean contains(int[] arr, int target) {
         for (int num : arr) {
@@ -602,7 +375,7 @@ public class FullFto {
 
     private int indexOfEdge(int target) {
         for (int i = 0; i < 12; i++) {
-            if (target == getEdge(i)) return i;
+            if (target == state.getEdge(i)) return i;
         }
         return -1;
     }
@@ -613,7 +386,7 @@ public class FullFto {
 
         int firstMatchingEdge = -1;
         for (int i = 0; i < 12; i++){
-            int edge = getEdge(i);
+            int edge = state.getEdge(i);
             if (contains(e, edge)){
                 firstMatchingEdge = edge;
                 break;
@@ -654,7 +427,7 @@ public class FullFto {
         int index = 0;
         int seen = 0;
         for (int i = 8; i >= 0; i--) {
-            int edge = getEdge(i);
+            int edge = state.getEdge(i);
             int smaller = Integer.bitCount(seen & ((1 << edge) - 1));
             index += smaller * FACTORIAL[8 - i];
             seen |= 1 << edge;
@@ -668,7 +441,7 @@ public class FullFto {
         long[] centerKeys = PHASE2_CENTER_KEYS[center.ordinal()];
 
         for (int i = 0; i < 24; i++) {
-            if (getCenterOrdinal(i) == centerId){
+            if (state.getCenterOrdinal(i) == centerId){
                 hash ^= centerKeys[i];
             }
         }
@@ -683,7 +456,7 @@ public class FullFto {
     public long phaseTwoHash(){
         long hash = 0;
 
-        hash ^= (centers & 0b111111111111111111000000000000000000000000L);
+        hash ^= (state.centers & 0b111111111111111111000000000000000000000000L);
 
         hash ^= edgeHash(CenterOrd.R);
         hash ^= edgeHash(CenterOrd.L);
@@ -696,12 +469,12 @@ public class FullFto {
         long hash = 0;
 
         for (int i = 0; i < 9; i++) {
-            hash ^= PHASE3_EDGE_KEYS[i][getEdge(i)];
+            hash ^= PHASE3_EDGE_KEYS[i][state.getEdge(i)];
         }
 
         for (int i = 0; i < 6; i++) {
-            int corner = getCorner(i);
-            hash ^= PHASE3_CORNER_KEYS[i][getCornerIndex(corner)][getCornerOrientation(corner)];
+            int corner = state.getCorner(i);
+            hash ^= PHASE3_CORNER_KEYS[i][state.getCornerIndex(corner)][state.getCornerOrientation(corner)];
         }
 
         return hash;
@@ -724,12 +497,12 @@ public class FullFto {
         long hash = 0;
         for (int cid = 0; cid < 6; cid++) {
             for (int cside = 0; cside < 2; cside++){
-                int corner = getCorner(cid);
-                int ci = getCornerIndex(corner);
-                int co = getCornerOrientation(corner);
+                int corner = state.getCorner(cid);
+                int ci = state.getCornerIndex(corner);
+                int co = state.getCornerOrientation(corner);
                 int matchingCenter = MATCHING_CENTER_INDICES[ci][(co + (2 * cside)) % 4];
                 for (long i = 0; i < 24; i++) {
-                    if (matchingCenter == getCenterIndex((int)i)){
+                    if (matchingCenter == state.getCenterIndex((int)i)){
                         hash |= i << (10 * cid + 5 * cside);
                     }
                 }
@@ -742,16 +515,16 @@ public class FullFto {
     public boolean checkPhaseTwoTripleData(long hash){
         for (int cid = 0; cid < 6; cid++) {
             for (int cside = 0; cside < 2; cside++){
-                int corner = getCorner(cid);
-                int ci = getCornerIndex(corner);
-                int co = getCornerOrientation(corner);
+                int corner = state.getCorner(cid);
+                int ci = state.getCornerIndex(corner);
+                int co = state.getCornerOrientation(corner);
 
                 long targetIndex = hash & 0b11111;
                 hash >>= 5;
 
                 int matchingCenter = MATCHING_CENTERS[ci][(co + (2 * cside)) % 4];
 
-                if (getCenter((int)targetIndex) != matchingCenter){
+                if (state.getCenter((int)targetIndex) != matchingCenter){
                     return false;
                 }
             }
@@ -882,42 +655,42 @@ public class FullFto {
 
         for (int i = 0; i < 6; i++) {
             int target = r.nextInt(6);
-            swapCorners(i, target);
+            state.swapCorners(i, target);
             if (i != target)
                 parity++;
         }
 
         if (parity % 2 == 1)
-            swapCorners(0, 1);
+            state.swapCorners(0, 1);
 
         parity = 0;
         for (int i = 0; i < 6; i++) {
-            int corner = getCorner(i);
-            if ((i < 3) == (getCornerIndex(corner) < 3)){
-                twistCorner(i, r.nextInt(2) * 2);
+            int corner = state.getCorner(i);
+            if ((i < 3) == (state.getCornerIndex(corner) < 3)){
+                state.twistCorner(i, r.nextInt(2) * 2);
             } else {
-                twistCorner(i, r.nextInt(2) * 2 + 1);
+                state.twistCorner(i, r.nextInt(2) * 2 + 1);
             }
-            parity += getCornerOrientation(getCorner(i));
+            parity += state.getCornerOrientation(state.getCorner(i));
         }
 
         if (parity % 4 == 2){
-            twistCorner(0, 2);
+            state.twistCorner(0, 2);
         }
 
         parity = 0;
         for (int i = 0; i < 12; i++) {
             int target = r.nextInt(12);
-            swapEdges(i, target);
+            state.swapEdges(i, target);
             if (target == i)
                 parity++;
         }
         if (parity % 2 == 1)
-            swapEdges(0, 1);
+            state.swapEdges(0, 1);
 
         for (int i = 0; i < 12; i++) {
-            swapCenters(i, r.nextInt(12));
-            swapCenters(i + 12, r.nextInt(12) + 12);
+            state.swapCenters(i, r.nextInt(12));
+            state.swapCenters(i + 12, r.nextInt(12) + 12);
         }
 
         clearMoveStack();
@@ -929,182 +702,30 @@ public class FullFto {
         return fto;
     }
 
+    private void turnState(Move move, InnerState s){
+
+    }
+
     public void turn(Move move){
         moveHistory.push(move);
-        cornerHistory.push(corners);
-        edgeHistory.push(edges);
-        centerHistory.push(centers);
+        stateHistory.push(new InnerState(state));
         if (trackCenterIndices) {
-            centerIndicesLowHistory.push(packedCenterIndicesLow);
-            centerIndicesHighHistory.push(packedCenterIndicesHigh);
+            centerIndicesLowHistory.push(state.packedCenterIndicesLow);
+            centerIndicesHighHistory.push(state.packedCenterIndicesHigh);
         }
-        switch (move){
-            case R:
-                cycleCorners(2, 1, 4);
-                twistCorner(2, 3);
-                twistCorner(1, 2);
-                twistCorner(4, 3);
-                cycleEdges(1, 5, 4);
-                cycleThreeCenters(15, 16, 17);
-                cycleThreeCenters(3, 1, 8);
-                cycleThreeCenters(4, 2, 6);
-                break;
-            case L:
-                cycleCorners(0, 2, 3);
-                twistCorner(0, 3);
-                twistCorner(2, 2);
-                twistCorner(3, 3);
-                cycleEdges(2, 3, 8);
-                cycleThreeCenters(12, 13, 14);
-                cycleThreeCenters(0, 3, 10);
-                cycleThreeCenters(2, 5, 9);
-                break;
-            case U:
-                cycleCorners(0, 1, 2);
-                cycleEdges(0, 1, 2);
-                cycleThreeCenters(0, 1, 2);
-                cycleThreeCenters(15, 12, 18);
-                cycleThreeCenters(16, 13, 19);
-                break;
-            case D:
-                cycleCorners(3, 4, 5);
-                cycleEdges(9, 10, 11);
-                cycleThreeCenters(21, 22, 23);
-                cycleThreeCenters(5, 8, 11);
-                cycleThreeCenters(4, 7, 10);
-                break;
-            case F:
-                cycleCorners(2, 4, 3);
-                twistCorner(2, 3);
-                twistCorner(4, 3);
-                twistCorner(3, 2);
-                cycleEdges(4, 9, 3);
-                cycleThreeCenters(3, 4, 5);
-                cycleThreeCenters(13, 17, 21);
-                cycleThreeCenters(15, 22, 14);
-                break;
-            case B:
-                cycleCorners(1, 0, 5);
-                twistCorner(1, 3);
-                twistCorner(0, 2);
-                twistCorner(5, 3);
-                cycleEdges(0, 6, 7);
-                cycleThreeCenters(18, 19, 20);
-                cycleThreeCenters(1, 9, 7);
-                cycleThreeCenters(0, 11, 6);
-                break;
-            case BR:
-                cycleCorners(1, 5, 4);
-                twistCorner(1, 3);
-                twistCorner(5, 3);
-                twistCorner(4, 2);
-                cycleEdges(5, 7, 10);
-                cycleThreeCenters(6, 7, 8);
-                cycleThreeCenters(16, 20, 22);
-                cycleThreeCenters(17, 18, 23);
-                break;
-            case BL:
-                cycleCorners(0, 3, 5);
-                twistCorner(0, 3);
-                twistCorner(3, 3);
-                twistCorner(5, 2);
-                cycleEdges(8, 11, 6);
-                cycleThreeCenters(9, 10, 11);
-                cycleThreeCenters(19, 14, 23);
-                cycleThreeCenters(12, 21, 20);
-                break;
-            case RP:
-                cycleCorners(2, 4, 1);
-                twistCorner(2, 2);
-                twistCorner(1, 1);
-                twistCorner(4, 1);
-                cycleEdges(1, 4, 5);
-                cycleThreeCenters(15, 17, 16);
-                cycleThreeCenters(3, 8, 1);
-                cycleThreeCenters(4, 6, 2);
-                break;
-            case LP:
-                cycleCorners(0, 3, 2);
-                twistCorner(0, 2);
-                twistCorner(2, 1);
-                twistCorner(3, 1);
-                cycleEdges(2, 8, 3);
-                cycleThreeCenters(12, 14, 13);
-                cycleThreeCenters(0, 10, 3);
-                cycleThreeCenters(2, 9, 5);
-                break;
-            case UP:
-                cycleCorners(0, 2, 1);
-                cycleEdges(0, 2, 1);
-                cycleThreeCenters(0, 2, 1);
-                cycleThreeCenters(15, 18, 12);
-                cycleThreeCenters(16, 19, 13);
-                break;
-            case DP:
-                cycleCorners(3, 5, 4);
-                cycleEdges(9, 11, 10);
-                cycleThreeCenters(21, 23, 22);
-                cycleThreeCenters(5, 11, 8);
-                cycleThreeCenters(4, 10, 7);
-                break;
-            case FP:
-                cycleCorners(2, 3, 4);
-                twistCorner(2, 1);
-                twistCorner(4, 2);
-                twistCorner(3, 1);
-                cycleEdges(4, 3, 9);
-                cycleThreeCenters(3, 5, 4);
-                cycleThreeCenters(13, 21, 17);
-                cycleThreeCenters(15, 14, 22);
-                break;
-            case BP:
-                cycleCorners(1, 5, 0);
-                twistCorner(1, 2);
-                twistCorner(0, 1);
-                twistCorner(5, 1);
-                cycleEdges(0, 7, 6);
-                cycleThreeCenters(18, 20, 19);
-                cycleThreeCenters(1, 7, 9);
-                cycleThreeCenters(0, 6, 11);
-                break;
-            case BRP:
-                cycleCorners(1, 4, 5);
-                twistCorner(1, 1);
-                twistCorner(5, 2);
-                twistCorner(4, 1);
-                cycleEdges(5, 10, 7);
-                cycleThreeCenters(6, 8, 7);
-                cycleThreeCenters(16, 22, 20);
-                cycleThreeCenters(17, 23, 18);
-                break;
-            case BLP:
-                cycleCorners(0, 5, 3);
-                twistCorner(0, 1);
-                twistCorner(3, 2);
-                twistCorner(5, 1);
-                cycleEdges(8, 6, 11);
-                cycleThreeCenters(9, 11, 10);
-                cycleThreeCenters(19, 23, 14);
-                cycleThreeCenters(12, 20, 21);
-                break;
-        }
+
+        state.turn(move);
     }
 
     public void undo(){
-        if (!cornerHistory.empty()) {
+        if (!moveHistory.empty()) {
             moveHistory.pop();
-            corners = cornerHistory.pop();
-            edges = edgeHistory.pop();
-            centers = centerHistory.pop();
-            if (trackCenterIndices) {
-                packedCenterIndicesLow = centerIndicesLowHistory.pop();
-                packedCenterIndicesHigh = centerIndicesHighHistory.pop();
-            }
+            state = stateHistory.pop();
         }
     }
 
     public boolean isSolved(){
-        return corners == SOLVED_CORNERS && edges == SOLVED_EDGES && centers == SOLVED_CENTERS;
+        return state.isSolved();
     }
 
     public void parseAlg(String alg){
